requirements.txt
numpy
scipy
matplotlib
skyfield
pyrtlsdr
PyQt5
pyqtgraph
